requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.1
attrs==20.3.0
CacheControl==0.12.6
cachetools==4.1.1
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.0
dj-database-url==0.5.0
Django==3.1.4
django-cors-headers==3.5.0
django-db-pooling==0.0.4
django-enumfield==2.0.1
django-heroku==0.3.1
django-restframework==0.0.1
djangorestframework==3.12.2
drf-yasg==1.17.0
elastic-apm==5.3.2
elasticsearch==7.1.0
elasticsearch-dsl==7.1.0
filelock==3.0.12
firebase-admin==4.4.0
google-api-core==1.23.0
google-api-python-client==1.12.8
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-cloud-core==1.4.4
google-cloud-firestore==2.0.2
google-cloud-storage==1.33.0
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
grpcio==1.34.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
importlib-metadata==3.1.1
inflection==0.3.1
iniconfig==1.1.1
isort==4.3.21
itypes==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.2.0
msgpack==1.0.0
packaging==20.7
Pillow==8.0.1
pluggy==0.13.1
proto-plus==1.13.0
protobuf==3.14.0
psycogreen==1.0.1
psycopg2==2.8.6
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.4.2
pylint==2.5.2
pyparsing==2.4.7
pytest==6.1.2
pytest-django==4.1.0
python-dateutil==2.8.1
python-memcached==1.59
pytz==2020.4
requests==2.25.0
rsa==4.6
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.2
virtualenv==20.0.21
wcwidth==0.1.8
whitenoise==5.2.0
wrapt==1.12.1
zipp==3.4.0
